(1803) ajouter la validation des financements
